+++ cpp/src/hdbscan/detail/stabilities.cuh
@@ -101,7 +101,7 @@ void compute_stabilities(const raft::handle_t& handle,
                              size_t&,
                              const value_t*,
                              value_t*,
-                             int,
+                             int64_t,
                              const value_idx*,
                              const value_idx*,
                              cudaStream_t) =
